Fix(audience): 공지 저장 뷰 이미지 오류 해결 (#339)
* fix: 저장한 공지 뷰 이미지 렌더링 수정

* refactor: 타입 정리

* refactor: type import 적용
