+++ apps/audience/src/shared/types/pagination.ts
@@ -1,0 +1,8 @@
+export interface Pagination {
+  currentPage: number;
+  totalPages: number;
+  totalElements: number;
+  size: number;
+  hasNext: boolean;
+  hasPrevious: boolean;
+}
